+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Update `yajra/laravel-datatables-html` to `~3.2`
-- Remove methods that already merged into the original yajra packages
+- Remove methods that have already merged into the original yajra packages
 
 ## 2.1.0 (2017-10-07)