+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Unreleased
+
+- Update `yajra/laravel-datatables-html` to `~3.2`
+- Remove methods that have already merged into the original yajra packages
+- Remove methods `stripedTable`, `hoveredTable` from Html Builder
+
 ## 2.1.0 (2017-10-07)
 
 - Automatic configure `datatables.engines.eloquent` and `datatables-buttons.stub`