+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Release Notes
 
-## Unreleased
+## 2.2.0 (2017-10-13)
 
 - Update `yajra/laravel-datatables-html` to `~3.2`
 - Remove methods that have already merged into the original yajra packages
 - Remove methods `stripedTable`, `hoveredTable` from Html Builder
 - Remove `addColumns` from `EloquentDataTable` since the original class has added https://github.com/yajra/laravel-datatables/pull/1416
+- Change default filename for exporting to "classname_Ymd_His"
 
 ## 2.1.0 (2017-10-07)